+++ docs/components/home/examples/ExDatePicker.vue
@@ -1,18 +1,22 @@
 <template>
   <v-date-picker
+    class='picker'
+    :from-page.sync='fromPage'
+    :to-page.sync='toPage'
     :mode='mode'
     :select-color='selectColor'
     :drag-color='dragColor'
     :disabled-dates='showDisabledDates ? disabledDates : null'
-    :attributes='[todayAttribute]'
+    :attributes='[]'
     :is-inline='isInline'
+    :is-expanded='isExpanded'
+    :popover-expanded='popoverExpanded'
     :popover-visibility='popoverVisibility'
     :popover-direction='popoverDirection'
     :popover-align='popoverAlign'
     inputClass='input'
     v-model='selectedValue'
-    is-double-paned
-    is-expanded>
+    is-double-paned>
   </v-date-picker>
 </template>
 
@@ -29,12 +33,15 @@ export default {
     showDisabledDates: Boolean,
     isInline: Boolean,
     isExpanded: Boolean,
+    popoverExpanded: Boolean,
     popoverVisibility: { type: String, default: 'hover' },
     popoverDirection: { type: String, default: 'bottom' },
     popoverAlign: { type: String, default: 'left' },
   },
   data() {
     return {
+      fromPage: null,
+      toPage: null,
       selectedValue: {
         start: new Date(thisMonthYear, thisMonth, 6),
         end: new Date(nextMonthYear, nextMonth, 25),
@@ -58,3 +65,8 @@ export default {
   },
 };
 </script>
+
+<style lang='sass'>
+.picker
+  min-width: 220px
+</style>